+++ training_templates/__init__.py
@@ -1,7 +1,1 @@
-from training_templates.scikit_learn_templates.base_templates import SkLearnHyperoptBase
-from training_templates.scikit_learn_templates.random_forest_hyperopt_template import (
-    RandomForestHyperoptTrainer,
-)
-from training_templates.scikit_learn_templates.xgboost_hyperopt_template import (
-    XGBoostHyperoptTrainer,
-)
+from training_templates.trainers.sklearn import SkLearnPipelineBase, SkLearnHyperoptTrainer
